Move OCLC parser to oclc_classify.py
